requirements.txt
coverage==7.7.0
grpcio==1.70.0
grpcio-tools==1.70.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
protobuf==5.29.3
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
setuptools==75.8.2